cloud(dashboard): move member key policy from sidebar to modal
